实现部分指令集及字节码解释器

+++ mini-jvm/src/main/java/org/yzpang/jvm/instructions/base/Index16Instruction.java
@@ -9,7 +9,7 @@ public class Index16Instruction extends NoOperandsInstruction{
     /**
      * 双字节操作数, 表示常量池索引
      */
-    protected short index;
+    protected int index;
 
     @Override
     public void fetchOperands(BytecodeReader reader) {
